requirements.txt
fastapi
uvicorn
feedparser
openai
pydantic
python-dotenv
